+++ src/Network/Network/ChangeLog.md
@@ -30,6 +30,7 @@
 * Added breaking change notification for `Get-AzFirewall`, `New-AzFirewall`, `Set-AzFirewall` and `New-AzFirewallHubIpAddress`
 
 ## Version 4.20.0
+* Fixed a bug that does not return HubIPAddresses' PrivateIPAddress during a Get-AzFirewall command 
 * Fixed a bug that removes existing resource tags during a Set-AzFirewallPolicy command 
 * Fixed required parameter `PrivateLinkResourceType` missing issue for the following cmdlets [#18655]
     - `Get-AzPrivateEndpointConnection`